Combining get_total_uncerts and get_total_uncerts_trial
